:white_check_mark: [#377] Test review response serializer/endpoint
